+++ tests/conftest.py
@@ -1,6 +1,8 @@
 """Pytest fixtures for SniTun."""
 
 import asyncio
+import asyncio.sslproto
+from asyncio.streams import StreamReader
 from collections.abc import AsyncGenerator, Generator
 from contextlib import asynccontextmanager
 from dataclasses import dataclass, field
@@ -12,6 +14,7 @@ import select
 import socket
 import ssl
 from threading import Thread
+from typing import Any, cast
 from unittest.mock import patch
 
 from aiohttp import web
@@ -19,10 +22,11 @@ import pytest
 from pytest_aiohttp import AiohttpServer
 import trustme
 
-from snitun.client.connector import Connector
+from snitun.client.connector import Connector, ConnectorHandler
 from snitun.multiplexer.channel import MultiplexerChannel
 from snitun.multiplexer.core import Multiplexer
 from snitun.multiplexer.crypto import CryptoTransport
+from snitun.multiplexer.transport import ChannelTransport
 from snitun.server.listener_peer import PeerListener
 from snitun.server.listener_sni import SNIProxy
 from snitun.server.peer import Peer
@@ -31,6 +35,8 @@ from snitun.utils.aiohttp_client import SniTunClientAioHttp
 from snitun.utils.asyncio import asyncio_timeout
 
 from .server.const_fernet import FERNET_TOKENS
+
+_LOGGER = logging.getLogger(__name__)
 
 logging.basicConfig(level=logging.DEBUG)
 
@@ -374,3 +380,154 @@ async def connector_missing_certificate(
 ) -> Connector:
     """Create a connector."""
     return snitun_client_aiohttp_missing_certificate._connector
+
+
+class BufferedStreamReaderProtocol(asyncio.StreamReaderProtocol):
+    """Buffered stream reader protocol."""
+
+    writer: asyncio.StreamWriter
+    reader: asyncio.StreamReader
+
+    def __init__(
+        self,
+        stream_reader: asyncio.StreamReader,
+        loop: asyncio.AbstractEventLoop,
+        **kwargs: Any,  # noqa: ANN401
+    ) -> None:
+        """Initialize buffered stream reader protocol."""
+        self.reader = stream_reader
+        self.loop = loop
+        super().__init__(stream_reader, loop=loop, **kwargs)
+
+    def connection_made(self, transport: asyncio.Transport) -> None:
+        """Handle connection made."""
+        _LOGGER.debug("BufferedStreamReaderProtocol.connection_made: %s", transport)
+        self.writer = asyncio.StreamWriter(transport, self, self.reader, self.loop)
+        self.transport = transport
+
+    def data_received(self, data: bytes) -> None:
+        """Handle data received."""
+        _LOGGER.debug("BufferedStreamReaderProtocol.data_received: %s", data)
+        super().data_received(data)
+
+    def connection_lost(self, exc: Exception | None) -> None:
+        """Handle Connection lost."""
+        _LOGGER.debug("BufferedStreamReaderProtocol.connection_lost: %s", exc)
+        super().connection_lost(exc)
+
+
+@dataclass
+class _ConnectorWithStreams:
+    """Connector with streams."""
+
+    connections: list[BufferedStreamReaderProtocol]
+    connector: Connector
+
+
+def make_snitun_connector(
+    ssl_context: ssl.SSLContext,
+    whitelist: bool = False,
+) -> _ConnectorWithStreams:
+    """Make connector."""
+    connections: list[BufferedStreamReaderProtocol] = []
+    loop = asyncio.get_running_loop()
+
+    def protocol_factory() -> BufferedStreamReaderProtocol:
+        reader = StreamReader(loop=loop)
+        protocol = BufferedStreamReaderProtocol(reader, loop=loop)
+        connections.append(protocol)
+        return protocol
+
+    return _ConnectorWithStreams(
+        connections,
+        Connector(
+            protocol_factory=protocol_factory,
+            ssl_context=ssl_context,
+            whitelist=whitelist,
+        ),
+    )
+
+
+@dataclass
+class TLSWrappedTransport:
+    """TLS wrapped transport."""
+
+    channel: MultiplexerChannel
+    protocol: BufferedStreamReaderProtocol
+    transport: ChannelTransport
+
+
+@dataclass
+class SNITunLoopback:
+    """Server Transport wrapper."""
+
+    client: TLSWrappedTransport
+    server: TLSWrappedTransport
+
+
+@pytest.fixture
+async def snitun_loopback(
+    multiplexer_client: Multiplexer,
+    multiplexer_server: Multiplexer,
+    client_ssl_context: ssl.SSLContext,
+    server_ssl_context: ssl.SSLContext,
+) -> SNITunLoopback:
+    """Make snitun server transport."""
+    connector_with_streams = make_snitun_connector(server_ssl_context, whitelist=False)
+    connector = connector_with_streams.connector
+    multiplexer_client._new_connections = connector.handler  # noqa: SLF001
+    connector_handler: ConnectorHandler | None = None
+
+    def save_connector_handler(
+        loop: asyncio.AbstractEventLoop,
+        multiplexer: Multiplexer,
+        channel: MultiplexerChannel,
+        transport: ChannelTransport,
+    ) -> ConnectorHandler:
+        nonlocal connector_handler
+        connector_handler = ConnectorHandler(loop, multiplexer, channel, transport)
+        return connector_handler
+
+    loop = asyncio.get_running_loop()
+
+    with patch("snitun.client.connector.ConnectorHandler", save_connector_handler):
+        server_channel = await multiplexer_server.create_channel(
+            IP_ADDR,
+            lambda _: None,
+        )
+        server_transport = ChannelTransport(server_channel, multiplexer_server)
+        server_transport.start_reader()
+        server_reader = asyncio.StreamReader(loop=loop)
+        server_protocol = BufferedStreamReaderProtocol(server_reader, loop=loop)
+        tls_wrapped_server_transport = await loop.start_tls(
+            server_transport,
+            server_protocol,
+            client_ssl_context,
+            server_side=False,
+        )
+        assert tls_wrapped_server_transport is not None, "Failed to start TLS"
+        server_protocol.connection_made(tls_wrapped_server_transport)
+
+    assert isinstance(connector_handler, ConnectorHandler)
+    handler = cast(ConnectorHandler, connector_handler)
+    client_channel = handler._channel  # noqa: SLF001
+    assert client_channel._pause_resume_reader_callback is not None  # noqa: SLF001
+    assert (
+        client_channel._pause_resume_reader_callback  # noqa: SLF001
+        == handler._pause_resume_reader_callback  # noqa: SLF001
+    )
+
+    assert connector_with_streams.connections
+    assert len(connector_with_streams.connections) == 1
+    client_protocol = connector_with_streams.connections[0]
+    client_transport = handler._transport  # noqa: SLF001
+    await asyncio.sleep(0.1)
+    assert client_protocol.writer
+    assert server_protocol.writer
+    assert isinstance(server_transport.get_protocol(), asyncio.sslproto.SSLProtocol)
+    assert isinstance(client_transport.get_protocol(), asyncio.sslproto.SSLProtocol)
+
+    return SNITunLoopback(
+        client=TLSWrappedTransport(client_channel, client_protocol, client_transport),
+        server=TLSWrappedTransport(server_channel, server_protocol, server_transport),
+    )
